Move metrics to separate file
